refactor: move SplashPage to features layer and persist GoRouter instance
