+++ lib/core/routes/router_config.dart
@@ -6,31 +6,43 @@ import '../../features/auth/login/presentation/viewmodels/auth_viewmodel.dart';
 import '../../features/auth/login/presentation/viewmodels/login_viewmodel.dart';
 import '../../features/home/presentation/pages/home_page.dart';
 import '../../features/home/presentation/viewmodels/home_viewmodel.dart';
+import '../../features/score/presentation/pages/score_page.dart';
+import '../../features/score/presentation/viewmodels/score_viewmodel.dart';
+import '../../features/splash/presentation/pages/splash_page.dart';
 import '../di/injector.dart';
 import '../logger/app_logger.dart';
 import '../services/navigation_service.dart';
 import 'app_routes.dart';
 
 class AppRouter {
-  static GoRouter get router => GoRouter(
-        navigatorKey: NavigationService.navigatorKey,
-        redirect: (context, state) => _redirect(context, state, injector()),
-        refreshListenable: injector<AuthViewModel>(),
-        initialLocation: AppRoutes.login,
-        routes: [
-          GoRoute(
-            path: AppRoutes.splash,
-            builder: (context, state) => const SplashPage(),
+  static final GoRouter router = GoRouter(
+    navigatorKey: NavigationService.navigatorKey,
+    redirect: (context, state) => _redirect(
+      context,
+      state,
+      injector(),
+    ),
+    refreshListenable: injector<AuthViewModel>(),
+    initialLocation: AppRoutes.login,
+    routes: [
+      GoRoute(
+        path: AppRoutes.splash,
+        builder: (context, state) => const SplashPage(),
+      ),
+      GoRoute(
+        path: AppRoutes.login,
+        pageBuilder: (context, state) => CustomTransitionPage(
+          key: state.pageKey,
+          child: LoginPage(
+            viewModel: injector<LoginViewModel>(),
           ),
-          GoRoute(
-            path: AppRoutes.login,
-            pageBuilder: (context, state) => CustomTransitionPage(
-              key: state.pageKey,
-              child: LoginPage(
-                viewModel: injector<LoginViewModel>(),
-              ),
-              transitionsBuilder:
-                  (context, animation, secondaryAnimation, child) {
+          transitionsBuilder:
+              (
+                context,
+                animation,
+                secondaryAnimation,
+                child,
+              ) {
                 return FadeTransition(
                   opacity: animation,
                   child: SlideTransition(
@@ -39,9 +51,7 @@ class AppRouter {
                         begin: const Offset(0.0, 0.3),
                         end: Offset.zero,
                       ).chain(
-                        CurveTween(
-                          curve: Curves.easeOutCubic,
-                        ),
+                        CurveTween(curve: Curves.easeOutCubic),
                       ),
                     ),
                     child: ScaleTransition(
@@ -50,9 +60,7 @@ class AppRouter {
                           begin: 0.8,
                           end: 1.0,
                         ).chain(
-                          CurveTween(
-                            curve: Curves.easeOutCubic,
-                          ),
+                          CurveTween(curve: Curves.easeOutCubic),
                         ),
                       ),
                       child: child,
@@ -60,16 +68,22 @@ class AppRouter {
                   ),
                 );
               },
-            ),
-          ),
-          GoRoute(
-            path: AppRoutes.home,
-            builder: (context, state) => HomePage(
-              viewModel: injector<HomeViewModel>(),
-            ),
-          ),
-        ],
-      );
+        ),
+      ),
+      GoRoute(
+        path: AppRoutes.home,
+        builder: (context, state) => HomePage(
+          viewModel: injector<HomeViewModel>(),
+        ),
+      ),
+      GoRoute(
+        path: AppRoutes.score,
+        builder: (context, state) => ScorePage(
+          viewModel: injector<ScoreViewModel>(),
+        ),
+      ),
+    ],
+  );
 
   static Future<String?> _redirect(
     BuildContext context,
@@ -97,18 +111,5 @@ class AppRouter {
       return AppRoutes.home;
     }
     return null;
-  }
-}
-
-class SplashPage extends StatelessWidget {
-  const SplashPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: CircularProgressIndicator(),
-      ),
-    );
   }
 }
